pokemon_type_gradio/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
gradio>=3.0
numpy
